Use u32 for vertex indices to avoid overflow. Fix ggx15 coin chute issue

+++ core/hw/pvr/ta_vtx.cpp
@@ -1256,7 +1256,7 @@ public:
 	__forceinline
 		static void AppendSpriteVertexA(TA_Sprite1A* sv)
 	{
-		u16* idx=vdrc.idx.Append(6);
+		u32* idx = vdrc.idx.Append(6);
 		u32 vbase=vdrc.verts.used();
 
 		idx[0]=vbase+0;
@@ -1450,7 +1450,7 @@ public:
 
 		//allocate storage for BG poly
 		vd_rc.global_param_op.Append();
-		u16* idx=vd_rc.idx.Append(4);
+		u32* idx = vd_rc.idx.Append(4);
 		int vbase=vd_rc.verts.used();
 
 		idx[0]=vbase+0;
